+++ packages/neuron-wallet/src/services/cells.ts
@@ -223,13 +223,29 @@ export default class CellsService {
     return cells
   }
 
-  public static async getDaoCells(walletId: string): Promise<Cell[]> {
+  public static async getDaoCells({ walletId, lockArgs }: { walletId: string; lockArgs?: string }): Promise<Cell[]> {
     const outputs: OutputEntity[] = await getConnection()
       .getRepository(OutputEntity)
       .createQueryBuilder('output')
       .leftJoinAndSelect('output.transaction', 'tx')
       .where(
-        `
+        lockArgs
+          ? `
+        output.daoData IS NOT NULL AND
+        output.lockArgs = :lockArgs AND
+        (
+          output.status = :liveStatus OR
+          output.status = :sentStatus OR
+          tx.status = :failedStatus OR
+          (
+            (
+              output.status = :deadStatus OR
+              output.status = :pendingStatus
+            ) AND
+            output.depositTxHash is not null
+          )
+        )`
+          : `
         output.daoData IS NOT NULL AND
         (
           output.status = :liveStatus OR
@@ -250,71 +266,7 @@ export default class CellsService {
         )`,
         {
           walletId,
-          liveStatus: OutputStatus.Live,
-          sentStatus: OutputStatus.Sent,
-          failedStatus: TransactionStatus.Failed,
-          deadStatus: OutputStatus.Dead,
-          pendingStatus: OutputStatus.Pending,
-        }
-      )
-      .orderBy(`CASE output.daoData WHEN '0x0000000000000000' THEN 1 ELSE 0 END`, 'ASC')
-      .addOrderBy('tx.timestamp', 'ASC')
-      .getMany()
-
-    const cells: Cell[] = outputs.map(output => {
-      const cell = output.toModel()
-      if (!output.depositTxHash) {
-        // if deposit cell, set depositInfo
-        cell.setDepositInfo({
-          txHash: output.transaction!.hash,
-          timestamp: output.transaction!.timestamp!,
-        })
-      } else {
-        // if not deposit cell, set withdrawInfo
-        const withdrawTx = output.transaction
-        cell.setWithdrawInfo({
-          txHash: withdrawTx!.hash,
-          timestamp: withdrawTx!.timestamp!,
-        })
-      }
-      return cell
-    })
-
-    await Promise.all([CellsService.addDepositInfo(cells), CellsService.addUnlockInfo(cells)])
-
-    return cells
-  }
-
-  public static async getMultisigDaoCells(multisigConfig: MultisigConfigModel): Promise<Cell[]> {
-    const multiSignBlake160 = Multisig.hash(
-      multisigConfig.blake160s,
-      multisigConfig.r,
-      multisigConfig.m,
-      multisigConfig.n
-    )
-
-    const outputs: OutputEntity[] = await getConnection()
-      .getRepository(OutputEntity)
-      .createQueryBuilder('output')
-      .leftJoinAndSelect('output.transaction', 'tx')
-      .where(
-        `
-        output.daoData IS NOT NULL AND
-        output.lockArgs = :multiSignBlake160 AND
-        (
-          output.status = :liveStatus OR
-          output.status = :sentStatus OR
-          tx.status = :failedStatus OR
-          (
-            (
-              output.status = :deadStatus OR
-              output.status = :pendingStatus
-            ) AND
-            output.depositTxHash is not null
-          )
-        )`,
-        {
-          multiSignBlake160,
+          lockArgs,
           liveStatus: OutputStatus.Live,
           sentStatus: OutputStatus.Sent,
           failedStatus: TransactionStatus.Failed,
